Working on distance-and-detection

+++ distance-and-detection/det_dist.cpp
@@ -148,7 +148,7 @@ int main(int argc, char * argv[]) try
 
         cv::Mat color_image_raw(cv::Size(color_width, color_height), CV_8UC3, colorData, cv::Mat::AUTO_STEP);
         cv::Mat colorMat;
-        cv::Mat rgbaMat(cv::Size(color_width, color_height), CV_32FC4, colorData_flt_CPU, cv::Mat::AUTO_STEP);
+        cv::Mat rgbaMat(cv::Size(color_width, color_height), CV_64FC4, colorData_flt_CPU, cv::Mat::AUTO_STEP);
         cv::cvtColor(color_image_raw, colorMat, cv::COLOR_RGB2BGR);
         cv::cvtColor(color_image_raw, rgbaMat, cv::COLOR_RGB2RGBA);
